refactor(tests): restructure test suite into modular architecture
- Each module now follows single-responsibility principle
- Improved maintainability and parallel test execution support
- Added comprehensive documentation for new test structure
- All 68 tests passing, no functionality loss
