+++ tests/cli/__init__.py
@@ -1,0 +1,1 @@
+"""CLI tests package."""
