+++ tests/large_diagrams/__init__.py
@@ -1,0 +1,1 @@
+"""Large diagram tests package."""
